+++ virtualFittingApp/src/shared/components/pagination/types/Pagination.d.ts
@@ -1,0 +1,8 @@
+import type { Dispatch, SetStateAction } from "react";
+
+export type PaginationType = {
+  page: number;
+  setPage: Dispatch<SetStateAction<number>>;
+  totalPage: number;
+  size: number;
+};
